moved logging layer
